requirements.txt
Django==4.0.5
environs==9.5.0
psycopg2-binary==2.9.3
Pillow==9.1.1
djangorestframework==3.13.1
gunicorn==20.1.0
pre-commit==2.19.0